all pages

+++ saas-service/src/main/java/com/example/medi/saas/dto/SaasStaffRequest.java
@@ -59,7 +59,7 @@ public class SaasStaffRequest {
 
     private BigDecimal consultationFee;
 
-    private BigDecimal onlineConsultationFee;
+    private Long onlineConsultationFee;
 
     private Boolean onlineConsultationEnabled;
 }
