+++ saas-service/src/main/java/com/example/medi/saas/entity/SaasStaff.java
@@ -93,7 +93,7 @@ public class SaasStaff {
 
     private BigDecimal consultationFee;
 
-    private BigDecimal onlineConsultationFee;
+    private Long onlineConsultationFee;
 
     private Boolean onlineConsultationEnabled = false;
